2019/06/15 1. HTTP服务端协议解析DateTime类型为时间戳 2. BuildLinkShell新增参数，可以控制显示或隐藏控制台

+++ EntryEngine/Utility.cs
@@ -1104,7 +1104,7 @@ namespace EntryEngine
 		{
 			return LocalTime + timestamp;
 		}
-        public static DateTime ToTime(int milliseconds)
+        public static DateTime ToTime(long milliseconds)
         {
             return LocalTime.AddMilliseconds(milliseconds);
         }
